pyproject.toml
[project]
name = "drug-kg-trial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-neo4j>=0.9.0",
    "langchain-openai==1.2.1",
    "neo4j>=6.1.0",
    "numpy>=2.4.4",
    "openai>=2.33.0",
    "orjson>=3.11.8",
    "pandas>=3.0.2",
    "pyarrow>=24.0.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.57.0",
    "tenacity>=9.1.4",
    "tqdm>=4.67.3",
    "uvicorn[standard]>=0.46.0",
]
